Add "find" command
This command takes a search term and returns any matching user, role,
game, or emote. It does not match channels, as it's not easy to guess
how secret a server's hidden channels may be from the average
non-API-savvy user.

I suspect this generates a lot of garbage but at the moment I don't
really think it will make any measurable impact on things.